Implement router allowance tracking

+++ app/src/main/java/com/crosschain/assettracker/ui/mvi/main/MainIntent.kt
@@ -1,11 +1,12 @@
 package com.crosschain.assettracker.ui.mvi.main
 
 import com.crosschain.assettracker.domain.model.Chain
+import java.math.BigInteger
 
 sealed class MainIntent {
     data class LoadBalance(val chain: Chain) : MainIntent()
     data class TrackTransfer(val messageId: String) : MainIntent()
-    data class TransferTokens(val amount: Long, val sourceChain: Chain, val destinationChain: Chain) : MainIntent()
+    data class TransferTokens(val amount: BigInteger, val sourceChain: Chain, val destinationChain: Chain) : MainIntent()
 
     data object LoadAccountFromWallet : MainIntent()
     data object LoadAccountFromCache : MainIntent()
